Move and extract things
