+++ docs/examples/TabSelector.tsx
